Make random test runnable.
Use listOf1 to avoid illegal regular expressions (regexp with two consecutive |),
at the price of not generating regular expressions that match empty (eg, "ab|").

Reduce alt depth to 1, which is way too little, because the test was taking way
too long to run. Now it takes just too long -- the whole "listOf1" thing needs to
be changed, to both enable greater depths and solve the previous problem.

Make subexpr always generate a non-empty string, again to solve the problem of
consecutive |, which is illegal.